outsource JPA query repository method parsing to different class
